requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
openai==1.6.1
python-dotenv==1.0.0
markdown==3.4.4
schedule==1.2.0
nltk==3.8.1
scikit-learn==1.3.0
pydantic==2.0.3
tabulate==0.9.0
wordcloud==1.9.2
emoji==2.8.0
pymongo==4.6.0
crawl4ai>=0.5.0 